+++ flake.nix
@@ -13,7 +13,7 @@
         let pkgs = nixpkgs.legacyPackages.${system}; in {
           default =
             let
-              appVersion = "0.2.0";
+              appVersion = "0.3.0";
               module = "github.com/tu-graz/kanboard-cli";
             in
             pkgs.buildGoModule {